+++ fletcher/_algorithms.py
@@ -5,7 +5,7 @@ import numpy as np
 import pyarrow as pa
 from pandas.core import nanops
 
-from fletcher._numba_compat import njit
+from fletcher._compat import njit
 from fletcher.algorithms.utils.chunking import dispatch_chunked_binary_map