+++ fletcher/algorithms/string.py
@@ -5,7 +5,7 @@ import numpy as np
 import pyarrow as pa
 
 from fletcher._algorithms import _buffer_to_view, _merge_valid_bitmaps
-from fletcher._numba_compat import njit
+from fletcher._compat import njit
 from fletcher.algorithms.utils.chunking import (
     _calculate_chunk_offsets,
     _combined_in_chunk_offsets,